```
goos: linux
goarch: arm64
pkg: github.com/puhitaku/go-waifu2x
name                old time/op    new time/op    delta
Waifu/Neko-4           1.87s ± 3%     1.36s ± 4%  -27.62%  (p=0.008 n=5+5)
Waifu/Wikipe-tan-4      193s ± 2%      134s ± 1%  -30.81%  (p=0.008 n=5+5)
```
please confirm the effect in your environment.