@orisano 2022-02-07
More improvements
```
Waifu/Wikipe-tan-4      193s ± 2%       87s ± 1%   -55.17%  (p=0.008 n=5+5)
```